Fix bugs.

+++ tasks/scaffolding-append.js
@@ -33,7 +33,7 @@ var task = {
                     if (opt.type === 'file') {
                         fs.readFile(opt.data[key], function (err, contents) {
                             if (err) {
-                                next(err);
+                                return next(err);
                             }
 
                             data[key] = contents + '{{' + key + '}}';
@@ -49,19 +49,30 @@ var task = {
                     async.forEach(files, function (file, next) {
                         glob(file, function (err, files) {
                             if (err) {
-                                next(err);
+                                return next(err);
                             }
 
                             // For each file in the glob result,
                             // perform the interpolation
                             async.forEach(files, function (file, next) {
-                                fs.readFile(file, function (err, contents) {
+                                // Only apply to files
+                                fs.stat(file, function (err, stat) {
                                     if (err) {
-                                        next(err);
+                                        return next(err);
                                     }
 
-                                    contents = stringLib.interpolate(contents.toString(), data);
-                                    fs.writeFile(file, contents, next);
+                                    if (!stat.isFile()) {
+                                        return next();
+                                    }
+
+                                    fs.readFile(file, function (err, contents) {
+                                        if (err) {
+                                            return next(err);
+                                        }
+
+                                        contents = stringLib.interpolate(contents.toString(), data);
+                                        fs.writeFile(file, contents, next);
+                                    });
                                 });
                             }, next);
                         });
